Fixed an accuracy calculation error in test_eval_gsm.py
After fine-tuning Aquila using the GSM8K training set, the output may contain "####". This causes the output answer to be treated as a reference answer, resulting in an increment of the accuracy variable and causing a calculation error in the final accuracy.
To resolve this issue, modify the code to only search for "####" in the "Reference Answer" section.